third_party/blink/web_tests/fast/events/click-over-descendant-elements-expected.txt

Check if any mousedown-mouseup pairs in one element dispatch click events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

This test needs to run in a test environment.
Very normal click on container:
PASS lastClickTarget is container
Move inside container:
PASS lastClickTarget is container
Move from container to a child:
PASS lastClickTarget is container
Move from a child to container:
PASS lastClickTarget is container
Move from a child to another child:
PASS lastClickTarget is container
Move out from a child:
PASS lastClickTarget is document.body
Removing a child element in the light DOM during a click:
PASS lastClickTarget is null
Click on escaping button content:
PASS lastClickTarget is button1
Click on disappearing INPUT value:
PASS lastClickTarget is input1
Mousedown on a form control, and mouseup on an element outside:
PASS lastClickTarget is document.body
PASS successfullyParsed is true

TEST COMPLETE
